fix: test-failure for struct alignment; memory leak in test suite

+++ test/include/test/test_log.h
@@ -23,6 +23,6 @@ void test_log_pass(TestLog* log, String message);
 
 FormattedOStream* get_fstream(TestLog* log);
 
-size_t summarize_tests(TestLog* log, Allocator* a);
+int summarize_tests(TestLog* log, Allocator* a);
 
 #endif
